fix: handling of pre-v0.1 when coercing implicit peer dependencies to minor (#38)
Previously, monoweave would throw an error if coerceImplicitPeerDependency was set to minor and a workspace peer dependency was pre-v0.1. This is because the version 0.0.x would get coerced to 0.0.0 resulting in the SemVer range of ^0.0.0. This range does not match anything, resulting in invalid peer dependencies. Related to https://github.com/yarnpkg/berry/issues/5841.