fix: do not reprocess version files
Remove any version files which have already been consumed. i.e. check if there are deletions on the remote. This is only necessary if running in an environment where we haven't pulled the latest changes on the main publish branch, e.g. GitHub's default behaviour when you don't override actions/checkout "ref" to point to the branch. This is best effort.